backend/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "mohican-backend"
version = "0.1.0"
description = "Backend job runner for Mohican red-team scans using promptfoo and garak."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi",
  "uvicorn",
  "pydantic>=2",
  "PyYAML",
]

[project.scripts]
mohican-backend = "mohican_backend.main:run"

[tool.setuptools.packages.find]
where = ["src"]

